requirements_lock.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe --no-emit-index-url --output-file=requirements_lock.txt requirements.txt
#
annotated-types==0.7.0
    # via pydantic
anthropic==0.85.0
    # via -r requirements.txt
anyio==4.12.1
    # via
    #   anthropic
    #   httpx
black==26.3.1
    # via -r requirements.txt
blinker==1.9.0
    # via flask
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   black
    #   flask
cryptography==46.0.5
    # via google-auth
distro==1.9.0
    # via anthropic
docstring-parser==0.17.0
    # via anthropic
flask==3.1.3
    # via -r requirements.txt
google-api-core==2.30.0
    # via google-api-python-client
google-api-python-client==2.192.0
    # via -r requirements.txt
google-auth==2.49.1
    # via
    #   -r requirements.txt
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.3.0
    # via
    #   -r requirements.txt
    #   google-api-python-client
google-auth-oauthlib==1.3.0
    # via -r requirements.txt
googleapis-common-protos==1.73.0
    # via google-api-core
grpcio==1.78.0
    # via
    #   -r requirements.txt
    #   grpcio-tools
grpcio-tools==1.62.3
    # via -r requirements.txt
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httplib2==0.31.2
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httpx==0.28.1
    # via anthropic
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
jiter==0.13.0
    # via anthropic
librt==0.8.1
    # via mypy
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
mypy==1.19.1
    # via -r requirements.txt
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
oauthlib==3.3.1
    # via requests-oauthlib
packaging==26.0
    # via
    #   black
    #   pytest
pathspec==1.0.4
    # via
    #   black
    #   mypy
platformdirs==4.9.4
    # via black
pluggy==1.6.0
    # via pytest
proto-plus==1.27.1
    # via google-api-core
protobuf==4.25.8
    # via
    #   -r requirements.txt
    #   google-api-core
    #   googleapis-common-protos
    #   grpcio-tools
    #   proto-plus
pyasn1==0.6.2
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via anthropic
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pyparsing==3.3.2
    # via httplib2
pytest==9.0.2
    # via -r requirements.txt
pytokens==0.4.1
    # via black
requests==2.32.5
    # via
    #   -r requirements.txt
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
sniffio==1.3.1
    # via anthropic
types-protobuf==6.32.1.20260221
    # via -r requirements.txt
types-requests==2.32.4.20260107
    # via -r requirements.txt
typing-extensions==4.15.0
    # via
    #   anthropic
    #   anyio
    #   grpcio
    #   mypy
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.6.3
    # via
    #   requests
    #   types-requests
werkzeug==3.1.6
    # via flask

# The following packages are considered to be unsafe in a requirements file:
setuptools==82.0.1
    # via grpcio-tools
